+++ resources/views/members/elements/premium/premium.blade.php
@@ -1,11 +1,13 @@
 @php use App\Models\UserRole; @endphp
 <div>
     <div class="p-4 rounded-1 bg-blue-700">
-        <h3 class="fw-normal fs-5 mb-0">
-            <span class='btn-role btn-premium rounded-1'><i class='me-2 {{ UserRole::ICON_PREMIUM }}'></i>Premium</span>
-        </h3>
-        <span class="home_abonnement_price fw-bold fs-2">14.99€</span>
-        <ul class="nav">
+        <div class="d-flex justify-content-between align-items-center">
+            <h3 class="fw-normal">
+                <span class='btn-role btn-premium rounded-1'><i class='me-2 {{ UserRole::ICON_PREMIUM }}'></i>Premium</span>
+            </h3>
+            <span class="home_abonnement_price fw-bold fs-2">14.99€</span>
+        </div>
+        <ul class="nav d-flex flex-column">
             <li class="py-1 d-flex text-green-light"><i class="bi bi-check-circle me-3"></i>{!! __('upgrade.roles.zmenu+') !!}</li>
             <li class="py-1 d-flex text-green-light"><i class="bi bi-check-circle me-3"></i>{{ __('upgrade.roles.creator_dashboard') }}</li>
             <li class="py-1 d-flex text-green-light"><i class="bi bi-check-circle me-3"></i>{{ __('upgrade.roles.discord_support_ticket') }}</li>
@@ -29,6 +31,17 @@
             <li class="py-1 d-flex text-muted"><i class="bi bi-check-circle me-3"></i>{{ __('upgrade.roles.auto_response') }}</li>
             <li class="py-1 d-flex text-muted"><i class="bi bi-check-circle me-3"></i>{{ __('upgrade.roles.username') }}</li>
         </ul>
-        <a href="{{ route('premium.checkout', UserRole::PREMIUM) }}" class="btn btn-success w-100 rounded-1 mt-4">{{ __('upgrade.purchase', ['price' => '14.99']) }}</a>
+        @auth()
+            @if(user()->role->isPremium())
+                <div class="btn btn-secondary w-100 rounded-1 mt-4 disabled">{{ __('upgrade.already') }}</div>
+            @else
+                <a href="{{ route('premium.checkout', UserRole::PREMIUM) }}"
+                   class="btn btn-success w-100 rounded-1 mt-4">{{ __('upgrade.purchase', ['price' => '14.99']) }}</a>
+            @endif
+        @endauth
+        @guest()
+            <a href="{{ route('premium.checkout', UserRole::PREMIUM) }}"
+               class="btn btn-success w-100 rounded-1 mt-4">{{ __('upgrade.purchase', ['price' => '14.99']) }}</a>
+        @endguest
     </div>
 </div>